# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateClientRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - address
    - email
    - name
    - phone
    type: object
...
  dto.UpdateClientRequest:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
    required:
    - id
    type: object
...
  utils.GenericResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for your invoice app.
  title: Smart Invoice Generator API
  version: "1.0"
paths:
...
  /v1/protected/clients:
    get:
      description: 'Retrieves all clients for the authenticated user with pagination
        (default: page=1, page_size=10)'
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search term for filtering clients
        in: query
        name: search
        type: string
      - description: Return all clients without pagination (use with caution)
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Get all clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Creates a new client for the authenticated user
      parameters:
      - description: Client data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Create a new client
      tags:
      - clients
  /v1/protected/clients/{id}:
    delete:
      description: Soft deletes a client by its ID for the authenticated user (can
        be restored)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Delete client
      tags:
      - clients
    get:
      description: Retrieves a client by its ID for the authenticated user
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Get client by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Updates a client by its ID for the authenticated user
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Update client
      tags:
      - clients
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"